+++ cffconvert/behavior_1_0_x/citation.py
@@ -8,6 +8,7 @@ from cffconvert.behavior_1_0_x.bibtex import BibtexObject
 from cffconvert.behavior_1_0_x.codemeta import CodemetaObject
 from cffconvert.behavior_1_0_x.endnote import EndnoteObject
 from cffconvert.behavior_1_0_x.schemaorg import SchemaorgObject
+from cffconvert.behavior_1_0_x.ris import RisObject
 from cffconvert.behavior_1_0_x.zenodo import ZenodoObject
 
 
@@ -50,6 +51,9 @@ class Citation_1_0_x(Contract):
 
     def as_schemaorg(self):
         return SchemaorgObject(self.cffobj).print()
+
+    def as_ris(self):
+        return RisObject(self.cffobj).print()
 
     def as_zenodo(self):
         return ZenodoObject(self.cffobj).print()
