+++ cffconvert/behavior_1_1_x/citation.py
@@ -8,6 +8,7 @@ from cffconvert.behavior_1_1_x.bibtex import BibtexObject
 from cffconvert.behavior_1_1_x.codemeta import CodemetaObject
 from cffconvert.behavior_1_1_x.endnote import EndnoteObject
 from cffconvert.behavior_1_1_x.schemaorg import SchemaorgObject
+from cffconvert.behavior_1_1_x.ris import RisObject
 from cffconvert.behavior_1_1_x.zenodo import ZenodoObject
 
 
@@ -48,6 +49,9 @@ class Citation_1_1_x(Contract):
 
     def as_schemaorg(self):
         return SchemaorgObject(self.cffobj).print()
+
+    def as_ris(self):
+        return RisObject(self.cffobj).print()
 
     def as_zenodo(self):
         return ZenodoObject(self.cffobj).print()
