+++ cffconvert/cli.py
@@ -1,7 +1,6 @@
 import sys
 import click
 from cffconvert.citation import Citation
-from cffconvert.ris import RisObject
 from cffconvert.version import __version__ as cffconvert_version
 
 
@@ -91,7 +90,7 @@ def cli(infile, outfile, outputformat, url, show_help, show_trace, validate, ign
     elif outputformat == "endnote":
         outstr = citation.as_endnote()
     elif outputformat == "ris":
-        outstr = RisObject(citation.cffobj).print()
+        outstr = citation.as_ris()
     elif outputformat == "schema.org":
         outstr = citation.as_schemaorg()
     elif outputformat == "zenodo":
