+++ test/contracts/RisObject.py
@@ -13,7 +13,7 @@ class Contract(ABC):
         pass
 
     @abstractmethod
-    def test_check_cff_object(self, fixture):
+    def test_check_cffobj(self, fixture):
         pass
 
     @abstractmethod
